Add complete open source project infrastructure (#13)
Add community health files, CI pipeline, release automation, and code
style enforcement to make Qtraktor a contributor-ready open source project.

Community:
- CONTRIBUTING.md with build, test, and PR guidelines
- CODE_OF_CONDUCT.md (Contributor Covenant v2.1)
- SECURITY.md with vulnerability reporting instructions
- YAML issue forms for bug reports and feature requests
- Pull request template with checklist

CI/CD:
- 3-platform CI workflow (Linux, macOS, Windows) with build and test
- clang-format lint with inline PR review comments (pinned to v18)
- clazy Qt-specific linter (warn-only, non-blocking)
- Conventional commit enforcement on PR titles
- Auto-labeling PRs by file path (11 labels)

Release automation:
- release-please for automated changelog and version management
- Seeded at v1.5.0 to match existing release history

Code style:
- .clang-format enforcing 4-space indent, Linux brace style
- .editorconfig for editor consistency
- .pre-commit-config.yaml for local format enforcement
- Normalize all source files to consistent style

Testing:
- Fuzz tests for .wpress parser (truncated, oversized, garbage,
  path traversal, boundary sizes, malformed JSON)
