+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,19 @@
+## What does this PR do?
+
+<!-- Brief description of the change -->
+
+## Related Issue
+
+<!-- Link to the issue this PR addresses, e.g., Fixes #123 -->
+
+## Checklist
+
+- [ ] PR title follows conventional commit format (`feat:`, `fix:`, `docs:`, etc.)
+- [ ] Code compiles without warnings on at least one platform
+- [ ] Tests pass locally (`cd tests && qmake tests.pro && make && ./tst_qtraktor`)
+- [ ] New code follows existing style (run `clang-format -i` on changed files)
+- [ ] Added tests for new functionality (if applicable)
+
+## Screenshots / Output
+
+<!-- If applicable, add screenshots or paste terminal output -->
