+++ .github/workflows/label.yml
@@ -1,0 +1,21 @@
+name: Label PRs
+
+on:
+  pull_request_target:
+    types: [opened, synchronize, reopened]
+
+# Security note: pull_request_target runs workflow code from the base branch with
+# write permissions. This is necessary for labeling PRs from forks. This workflow
+# must NEVER include a checkout step or run any user-supplied code.
+
+jobs:
+  label:
+    runs-on: ubuntu-latest
+    permissions:
+      contents: read
+      pull-requests: write
+    steps:
+      - uses: actions/labeler@v5
+        with:
+          repo-token: ${{ secrets.GITHUB_TOKEN }}
+          sync-labels: false
