+++ CODE_OF_CONDUCT.md
@@ -1,0 +1,9 @@
+# Code of Conduct
+
+This project follows the [Contributor Covenant v2.1](https://www.contributor-covenant.org/version/2/1/code_of_conduct/).
+
+## Enforcement
+
+Report unacceptable behavior to the project maintainers via GitHub issues or email.
+
+See the full text at https://www.contributor-covenant.org/version/2/1/code_of_conduct/
