--- mainwindow.cpp
@@ -1,286 +1,0 @@
-#include "mainwindow.h"
-#include "ui_mainwindow.h"
-#include <QFileDialog>
-#include <QIODevice>
-#include <QMessageBox>
-#include <QPushButton>
-#include <QProcess>
-#include <QSettings>
-#include "passworddialog.h"
-#include "cryptoutils.h"
-
-MainWindow::MainWindow(QWidget *parent) :
-  QMainWindow(parent),
-  ui(new Ui::MainWindow)
-{
-  ui->setupUi(this);
-  ui->progressBar->setVisible(false);
-  ui->logTextEdit->setVisible(false);
-  connect(ui->dropZone, &DropOverlay::fileDropped, this, &MainWindow::openBackupFile);
-  connect(ui->dropZone, &DropOverlay::clicked, this, &MainWindow::openBackup);
-  ui->clearButton->setVisible(false);
-
-  QSettings settings("com.servmask", "Traktor");
-  restoreGeometry(settings.value("windowGeometry").toByteArray());
-}
-
-MainWindow::~MainWindow()
-{
-  QSettings settings("com.servmask", "Traktor");
-  settings.setValue("windowGeometry", saveGeometry());
-  delete ui;
-}
-
-void MainWindow::openBackup()
-{
-  QSettings settings("com.servmask", "Traktor");
-  QString lastDir = settings.value("lastOpenPath").toString();
-
-  QString selectedFile = QFileDialog::getOpenFileName(
-    this,
-    tr("Open a backup"),
-    lastDir,
-    tr("WordPress backup (*.wpress)")
-  );
-
-  if (selectedFile.isNull()) {
-    return;
-  }
-
-  settings.setValue("lastOpenPath", QFileInfo(selectedFile).absolutePath());
-  backupFilename = selectedFile;
-
-  QFileInfo fileInfo(backupFilename);
-
-  if (!fileInfo.isReadable()) {
-    QMessageBox::warning(
-      this,
-      tr("Unable to open file"),
-      tr("Unable to open file: %1").arg(backupFilename),
-      QMessageBox::StandardButton::Ok
-    );
-    return;
-  }
-
-  openBackupFile(backupFilename);
-  filePassword.clear(); // Reset password when opening new file
-}
-
-void MainWindow::clearFile()
-{
-  backupFilename.clear();
-  filePassword.clear();
-  ui->dropZone->setFileName(QString());
-  ui->extractBackupButton->setEnabled(false);
-  ui->clearButton->setVisible(false);
-  ui->progressBar->setVisible(false);
-  ui->logTextEdit->setVisible(false);
-}
-
-void MainWindow::extractTo()
-{
-  QSettings settings("com.servmask", "Traktor");
-  QString lastDir = settings.value("lastExtractPath").toString();
-
-  QString extractToDir = QFileDialog::getExistingDirectory(
-    this,
-    tr("Select extract to folder"),
-    lastDir
-  );
-
-  if (extractToDir.isNull()) {
-    return;
-  }
-
-  settings.setValue("lastExtractPath", extractToDir);
-  extractToPath(extractToDir);
-}
-
-void MainWindow::setPassword(const QString &password)
-{
-  filePassword = password;
-}
-
-void MainWindow::extractToPath(const QString &destDir)
-{
-  QFileInfo fileInfo(backupFilename);
-  QDir extractTo(destDir + "/" + fileInfo.baseName());
-
-  if (extractTo.exists()) {
-    QMessageBox msgBox(this);
-    msgBox.setWindowTitle(tr("Directory already exists"));
-    msgBox.setText(tr("The directory %1 already exists.").arg(extractTo.path()));
-    QPushButton *wipeBtn = msgBox.addButton(tr("Wipe && Extract"), QMessageBox::DestructiveRole);
-    QPushButton *newBtn = msgBox.addButton(tr("Create New Folder"), QMessageBox::ActionRole);
-    msgBox.addButton(QMessageBox::Cancel);
-    msgBox.exec();
-
-    if (msgBox.clickedButton() == wipeBtn) {
-      if (!extractTo.removeRecursively()) {
-        QMessageBox::warning(this, tr("Unable to remove directory"),
-          tr("Unable to remove directory %1. Fix permissions and try again.").arg(extractTo.path()),
-          QMessageBox::StandardButton::Ok);
-        return;
-      }
-    } else if (msgBox.clickedButton() == newBtn) {
-      QString basePath = destDir + "/" + fileInfo.baseName();
-      int suffix = 1;
-      while (QDir(basePath + " (" + QString::number(suffix) + ")").exists() && suffix <= 100) {
-        suffix++;
-      }
-      if (suffix > 100) {
-        QMessageBox::warning(this, tr("Too many directories"),
-          tr("Too many directories with the name %1. Remove some and try again.").arg(fileInfo.baseName()),
-          QMessageBox::StandardButton::Ok);
-        return;
-      }
-      extractTo = QDir(basePath + " (" + QString::number(suffix) + ")");
-    } else {
-      return;
-    }
-  }
-
-  if (!QDir().mkdir(extractTo.path())) {
-     QMessageBox::warning(
-      this,
-      tr("Unable to create directory"),
-      tr("Unable to create directory %1. Fix permissions and try again.").arg(extractTo.path()),
-      QMessageBox::StandardButton::Ok
-    );
-    return;
-  }
-
-  // Quick config check on main thread to determine if password is needed
-  CheckResult config = ExtractionWorker::checkConfig(backupFilename);
-
-  if (!config.isValid) {
-    QMessageBox::warning(this, tr("Corrupted backup file"),
-      tr("The backup file is corrupted. It is missing the end of the file."),
-      QMessageBox::StandardButton::Ok);
-    extractTo.removeRecursively();
-    return;
-  }
-
-  if (config.isEncrypted && filePassword.isEmpty()) {
-    PasswordDialog dialog(this);
-    dialog.setWindowTitle(tr("Password Required"));
-
-    if (dialog.exec() == QDialog::Accepted) {
-      filePassword = dialog.getPassword();
-    } else {
-      extractTo.removeRecursively();
-      return;
-    }
-  }
-
-  currentExtractDir = extractTo.path();
-
-  ui->progressBar->setVisible(true);
-  ui->progressBar->setValue(0);
-  ui->logTextEdit->clear();
-  ui->logTextEdit->setVisible(false);
-
-  // Disable controls during extraction
-  ui->openBackupButton->setEnabled(false);
-  ui->extractBackupButton->setEnabled(false);
-
-  activeWorker = new ExtractionWorker(backupFilename, filePassword, currentExtractDir, this);
-
-  connect(activeWorker, &ExtractionWorker::progress, this, &MainWindow::extractProgress);
-  connect(activeWorker, &ExtractionWorker::extractionError, this, &MainWindow::onExtractionError);
-  connect(activeWorker, &ExtractionWorker::extractionFinished, this, &MainWindow::onExtractionFinished);
-  connect(activeWorker, &ExtractionWorker::logMessage, this, [this](const QString &msg) {
-    ui->logTextEdit->setVisible(true);
-    ui->logTextEdit->append(msg);
-  });
-  connect(activeWorker, &QThread::finished, activeWorker, &QObject::deleteLater);
-
-  activeWorker->start();
-}
-
-void MainWindow::onExtractionError(const QString &error)
-{
-  lastExtractionError = error;
-}
-
-void MainWindow::onExtractionFinished(bool success)
-{
-  activeWorker = nullptr;
-  ui->openBackupButton->setEnabled(true);
-  ui->progressBar->setVisible(false);
-
-  if (!success) {
-    ui->extractBackupButton->setEnabled(true);
-    filePassword.clear();
-    QDir(currentExtractDir).removeRecursively();
-
-    QString errorMessage;
-    if (!lastExtractionError.isEmpty()) {
-      errorMessage = lastExtractionError;
-    } else {
-      errorMessage = tr("The backup file extraction failed. The file may be corrupted or the password may be incorrect.");
-    }
-    QMessageBox msgBox(this);
-    msgBox.setWindowTitle(tr("Extraction failed"));
-    msgBox.setIcon(QMessageBox::Warning);
-    msgBox.setTextFormat(Qt::RichText);
-    msgBox.setTextInteractionFlags(Qt::TextBrowserInteraction);
-    msgBox.setText(errorMessage);
-    msgBox.setStandardButtons(QMessageBox::Ok);
-    msgBox.exec();
-  } else {
-    ui->dropZone->setFileName(tr("Extracted backup in %1").arg(currentExtractDir));
-    ui->extractBackupButton->setDisabled(true);
-    showInGraphicalShell(currentExtractDir);
-  }
-
-  lastExtractionError.clear();
-}
-
-void MainWindow::extractProgress(float percent)
-{
-  ui->progressBar->setValue(static_cast<int>(percent));
-}
-
-// copied form https://github.com/qt-creator/qt-creator/blob/master/src/plugins/coreplugin/fileutils.cpp#L67
-void MainWindow::showInGraphicalShell(const QString &pathIn)
-{
-    const QFileInfo fileInfo(pathIn);
-
-#if defined (Q_OS_WIN)
-  QStringList param;
-  if (!fileInfo.isDir())
-    param += QLatin1String("/select,");
-  param += QDir::toNativeSeparators(fileInfo.canonicalFilePath());
-  QProcess::startDetached("explorer", param);
-#endif
-
-#if defined (Q_OS_MAC)
-  QStringList scriptArgs;
-  scriptArgs << QLatin1String("-e")
-      << QString::fromLatin1("tell application \"Finder\" to reveal POSIX file \"%1\"")
-      .arg(fileInfo.canonicalFilePath());
-  QProcess::execute(QLatin1String("/usr/bin/osascript"), scriptArgs);
-  scriptArgs.clear();
-  scriptArgs << QLatin1String("-e")
-      << QLatin1String("tell application \"Finder\" to activate");
-  QProcess::execute(QLatin1String("/usr/bin/osascript"), scriptArgs);
-#endif
-}
-
-void MainWindow::openBackupFile(const QString &filename)
-{
-    backupFilename = filename;
-    QFileInfo fileInfo(backupFilename);
-
-    if (!fileInfo.isReadable()) {
-        QMessageBox::warning(this, tr("Unable to open file"),
-                           tr("Unable to open file: %1").arg(backupFilename),
-                           QMessageBox::StandardButton::Ok);
-        return;
-    }
-
-    ui->dropZone->setFileName(fileInfo.fileName());
-    ui->extractBackupButton->setEnabled(true);
-    ui->clearButton->setVisible(true);
-}
