+++ src/dockprogress.mm
