+++ src/dockprogress_stub.cpp
