+++ src/dropoverlay.h
@@ -7,23 +7,23 @@ class DropOverlay : public QWidget
 {
     Q_OBJECT
 
-  public:
+public:
     explicit DropOverlay(QWidget *parent = nullptr);
     void setHighlighted(bool highlighted);
     void setFileName(const QString &name);
 
-  signals:
+signals:
     void fileDropped(const QString &filePath);
     void clicked();
 
-  protected:
+protected:
     void paintEvent(QPaintEvent *event) override;
     void dragEnterEvent(QDragEnterEvent *event) override;
     void dragLeaveEvent(QDragLeaveEvent *event) override;
     void dropEvent(QDropEvent *event) override;
     void mousePressEvent(QMouseEvent *event) override;
 
-  private:
+private:
     bool m_highlighted = false;
     QString m_fileName;
 };
